# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "ATOS", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/activities/{activity_id}/shoe": {
      "put": {
        "tags": ["activities"],
        "summary": "Assign Activity Shoe",
        "description": "Asigna o desasigna una zapatilla a una actividad cardio (running).",
        "operationId": "activities-assign_activity_shoe",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Activity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityShoeAssignment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/shoes/": {
      "get": {
        "tags": ["shoes"],
        "summary": "Read Shoes",
        "operationId": "shoes-read_shoes",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": ["training", "race", "trail", "easy", "mixed"],
                  "type": "string"
                },
                { "type": "null" }
              ],
              "title": "Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShoesPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["shoes"],
        "summary": "Create Shoe",
        "operationId": "shoes-create_shoe",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ShoeCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShoePublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/shoes/import-strava": {
      "post": {
        "tags": ["shoes"],
        "summary": "Import Strava Shoes",
        "description": "Importa las zapatillas guardadas en Strava como calzado local.",
        "operationId": "shoes-import_strava_shoes",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShoeImportResult" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/shoes/{shoe_id}": {
      "get": {
        "tags": ["shoes"],
        "summary": "Read Shoe",
        "operationId": "shoes-read_shoe",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "shoe_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Shoe Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShoePublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["shoes"],
        "summary": "Update Shoe",
        "operationId": "shoes-update_shoe",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "shoe_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Shoe Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ShoeUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShoePublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["shoes"],
        "summary": "Delete Shoe",
        "operationId": "shoes-delete_shoe",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "shoe_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Shoe Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "string" },
                  "title": "Response Shoes-Delete Shoe"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/shoes/{shoe_id}/stats": {
      "get": {
        "tags": ["shoes"],
        "summary": "Read Shoe Stats",
        "operationId": "shoes-read_shoe_stats",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "shoe_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Shoe Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShoeStatsPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/shoes/{shoe_id}/activities": {
      "get": {
        "tags": ["shoes"],
        "summary": "Read Shoe Activities",
        "operationId": "shoes-read_shoe_activities",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "shoe_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Shoe Id" }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoeActivitiesPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/storage/upload": {
      "post": {
        "tags": ["storage"],
        "summary": "Upload File",
        "operationId": "storage-upload_file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_storage-upload_file"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": { "type": "string" },
                  "type": "object",
                  "title": "Response Storage-Upload File"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/storage/{filename}": {
      "get": {
        "tags": ["storage"],
        "summary": "Read File",
        "operationId": "storage-read_file",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Filename" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityCardioBase": {
        "properties": {
          "distance_meters": {
            "type": "number",
            "minimum": 0.0,
            "title": "Distance Meters",
            "default": 0
          },
          "avg_pace_seconds_per_km": {
            "anyOf": [{ "type": "number", "minimum": 0.0 }, { "type": "null" }],
            "title": "Avg Pace Seconds Per Km"
          },
          "avg_hr": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Avg Hr"
          },
          "max_hr": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Max Hr"
          },
          "elevation_gain_meters": {
            "type": "number",
            "minimum": 0.0,
            "title": "Elevation Gain Meters",
            "default": 0
          },
          "calories": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Calories"
          },
          "map_summary_polyline": {
            "anyOf": [
              { "type": "string", "maxLength": 20000 },
              { "type": "null" }
            ],
            "title": "Map Summary Polyline"
          },
          "splits": {
            "items": { "additionalProperties": true, "type": "object" },
            "type": "array",
            "title": "Splits"
          },
          "heart_rate_zones": {
            "items": { "additionalProperties": true, "type": "object" },
            "type": "array",
            "title": "Heart Rate Zones"
          },
          "shoe_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Shoe Id"
          },
          "gear_id": {
            "anyOf": [
              { "type": "string", "maxLength": 64 },
              { "type": "null" }
            ],
            "title": "Gear Id"
          }
        },
        "type": "object",
        "title": "ActivityCardioBase"
      },
...
      "ActivityPublic": {
        "properties": {
          "source_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Source Id"
          },
          "source_type": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Source Type"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "duration_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duration Seconds",
            "default": 0
          },
          "name": {
            "anyOf": [
              { "type": "string", "maxLength": 255 },
              { "type": "null" }
            ],
            "title": "Name"
          },
          "sport_type": {
            "anyOf": [
              { "type": "string", "maxLength": 32 },
              { "type": "null" }
            ],
            "title": "Sport Type"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "user_id": { "type": "string", "format": "uuid", "title": "User Id" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "cardio": {
            "anyOf": [
              { "$ref": "#/components/schemas/ActivityCardioBase" },
              { "type": "null" }
            ]
          },
          "strength": {
            "anyOf": [
              { "$ref": "#/components/schemas/ActivityStrengthBase" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "source_type",
          "timestamp",
          "id",
          "user_id",
          "created_at"
        ],
        "title": "ActivityPublic"
      },
      "ActivityShoeAssignment": {
        "properties": {
          "shoe_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Shoe Id"
          }
        },
        "type": "object",
        "title": "ActivityShoeAssignment"
      },
...
      "Body_storage-upload_file": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": ["file"],
        "title": "Body_storage-upload_file"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShoeActivitiesPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/ActivityPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "ShoeActivitiesPublic"
      },
      "ShoeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "brand": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Brand"
          },
          "model": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Model"
          },
          "category": {
            "type": "string",
            "enum": ["training", "race", "trail", "easy", "mixed"],
            "title": "Category"
          },
          "purchase_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "Purchase Date"
          },
          "target_distance_km": {
            "anyOf": [{ "type": "number", "minimum": 0.0 }, { "type": "null" }],
            "title": "Target Distance Km"
          },
          "color": {
            "anyOf": [
              { "type": "string", "maxLength": 50 },
              { "type": "null" }
            ],
            "title": "Color"
          },
          "photo_url": {
            "anyOf": [
              { "type": "string", "maxLength": 500 },
              { "type": "null" }
            ],
            "title": "Photo Url"
          },
          "notes": {
            "anyOf": [
              { "type": "string", "maxLength": 4000 },
              { "type": "null" }
            ],
            "title": "Notes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": ["name", "category"],
        "title": "ShoeCreate"
      },
      "ShoeImportResult": {
        "properties": {
          "imported": { "type": "integer", "title": "Imported", "default": 0 },
          "updated": { "type": "integer", "title": "Updated", "default": 0 },
          "linked": { "type": "integer", "title": "Linked", "default": 0 },
          "assigned": { "type": "integer", "title": "Assigned", "default": 0 },
          "message": { "type": "string", "title": "Message", "default": "" }
        },
        "type": "object",
        "title": "ShoeImportResult"
      },
      "ShoePublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "brand": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Brand"
          },
          "model": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Model"
          },
          "category": {
            "type": "string",
            "enum": ["training", "race", "trail", "easy", "mixed"],
            "title": "Category"
          },
          "purchase_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "Purchase Date"
          },
          "target_distance_km": {
            "anyOf": [{ "type": "number", "minimum": 0.0 }, { "type": "null" }],
            "title": "Target Distance Km"
          },
          "color": {
            "anyOf": [
              { "type": "string", "maxLength": 50 },
              { "type": "null" }
            ],
            "title": "Color"
          },
          "photo_url": {
            "anyOf": [
              { "type": "string", "maxLength": 500 },
              { "type": "null" }
            ],
            "title": "Photo Url"
          },
          "notes": {
            "anyOf": [
              { "type": "string", "maxLength": 4000 },
              { "type": "null" }
            ],
            "title": "Notes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "user_id": { "type": "string", "format": "uuid", "title": "User Id" },
          "strava_gear_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Strava Gear Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "ShoePublic"
      },
      "ShoeStatsPublic": {
        "properties": {
          "total_distance_meters": {
            "type": "number",
            "title": "Total Distance Meters",
            "default": 0
          },
          "sessions": { "type": "integer", "title": "Sessions", "default": 0 },
          "avg_pace_seconds_per_km": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Avg Pace Seconds Per Km"
          },
          "avg_hr": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Avg Hr"
          },
          "target_distance_km": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Target Distance Km"
          },
          "races": {
            "items": { "$ref": "#/components/schemas/RacePublic" },
            "type": "array",
            "title": "Races"
          }
        },
        "type": "object",
        "title": "ShoeStatsPublic"
      },
      "ShoeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              { "type": "string", "maxLength": 255, "minLength": 1 },
              { "type": "null" }
            ],
            "title": "Name"
          },
          "brand": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Brand"
          },
          "model": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Model"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["training", "race", "trail", "easy", "mixed"]
              },
              { "type": "null" }
            ],
            "title": "Category"
          },
          "purchase_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "Purchase Date"
          },
          "target_distance_km": {
            "anyOf": [{ "type": "number", "minimum": 0.0 }, { "type": "null" }],
            "title": "Target Distance Km"
          },
          "color": {
            "anyOf": [
              { "type": "string", "maxLength": 50 },
              { "type": "null" }
            ],
            "title": "Color"
          },
          "photo_url": {
            "anyOf": [
              { "type": "string", "maxLength": 500 },
              { "type": "null" }
            ],
            "title": "Photo Url"
          },
          "notes": {
            "anyOf": [
              { "type": "string", "maxLength": 4000 },
              { "type": "null" }
            ],
            "title": "Notes"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "ShoeUpdate"
      },
      "ShoesPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/ShoePublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "ShoesPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}